The rate limit is a property of the postMessage boundary, not of the loop: the throttle moves to the worker, and the module splits by altitude
`rankCandidatesInBrowser` wrapped its own `onProgress` in `throttleProgress`, justified as "a second
driver over this enumeration would inherit the throttle for free". That is backwards: 100 ms is a
TRANSPORT rate, chosen for postMessage + React, and a second consumer (a test rig, a node driver, a
future ETA estimator) inherited a ceiling it never asked for and could not observe every candidate
through. The driver now calls `onProgress` per candidate — a closure call against 25-167 ms of wasm
each, free — and `rank.worker.ts` throttles at the boundary the policy is about. Same ~10 msg/s
ceiling for every fan shape, one fresh throttle per request.

`rank-progress.ts` also fused three altitudes: the domain (phases), the transport (the throttle) and
the presentation (`progressLabel`'s user-facing sentences, `progressBar`'s pixel clamp) — so the
compute driver's import graph reached the UI's copy. Split: `rank-progress.ts` keeps the model and
the throttle (imported by score-wasm.ts and the worker), `progress-view.ts` takes the sentences and
the bar (imported by RankPanel.tsx), and the test file splits on the same seam.
